refact(Members): reorganize files
