+++ src/services/members/api.js
@@ -1,4 +1,4 @@
-import { prepareURL, getApi, putApi } from './_helpers';
+import { prepareURL, getApi, putApi } from 'api/_helpers.js';
 
 const membersUrl               = (orgId) => prepareURL(`/api/organizations/${orgId}/members`);
 const organizationLastVisitUrl = (orgId) => prepareURL(`/api/organizations/${orgId}/last_visit`);
