+++ src/services/members/selectors.js
@@ -1,0 +1,4 @@
+const selectMembers       = (state) => state.members.items;
+const selectMembersStatus = (state) => state.members.status;
+
+export { selectMembers, selectMembersStatus };
